+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.5.1
+
+* [FIXED] frozen CLI when we execute `integrity` cli command if no app is connected
+* [FIXED] `req` & `res` objects memory leak. Incorrect `req` & `res` objects were being provided to a `Route.prototype.catch` error handler function
+
+
 ## v0.5.0
 
 * [FIXED] min required `pg` module version (if installed as peer dependency)